api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.8.2
PyYAML==6.0.2
psycopg2-binary==2.9.9
